pyproject.toml
[project]
name = "ml-setup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kaggle>=2.0.0",
    "mlflow>=1.27.0",
    "notebook>=7.5.4",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "xgboost>=3.2.0",
]
